test(e2e): refactor commands
