+++ tests/e2e/specs/receive_transaction.spec.js
@@ -1,6 +1,6 @@
 describe('Receive Page', () => {
   beforeEach(() => {
-    cy.getInitialData();
+    cy.mockInitialData();
     cy.visit('#/receive');
     cy.mockWeb3Requests();
     cy.waitPageLoad();
